Added support for signing scripts and executable.
Added in functionality to support the automatic signing of all scripts, either the primary script or supplementary scripts, as well as the finalized output executable. Also changed all PowerShell.exe references to have the -NoProfile switch to speed up the finalized executable if the user has a PowerShell profile enabled. Updated readme to include additional parameters.